move test to use ktor assert
